WI-741: an uninstantiated spec type parameter is not a column type
A relation column typed only by a spec operation took that spec's own
parameter verbatim -- `rule named(?x) :- eq(?x, "root")` records `?x` at
`anthill.prelude.PartialEq.T` -- so a second, concretely-typed clause was
reported disjoint against it and the relation failed to LOAD, though both
clauses plainly meant String.

The ticket's hypothesis (resolve `var_types` through the substitution
`collect_rule_var_types` drops) was falsified by instrumenting the collector:
for `eq(?x, "root")` that substitution is EMPTY, since a literal argument is
not a var and constrains nothing. Where it is non-empty it binds the ONE
alias every `PartialEq.T` shares, which would have typed an Int64 variable
String.

Two producer-side fixes:

* `relation_clause_columns` normalizes a bare parameter into the raw
  `Var::Global` spelling, so "unknown" stays spelled ONE way and
  `join_column_types` keeps a single recognizer. ONE var per parameter
  SYMBOL, not per column -- the parameter says nothing about a column's type
  but everything about which columns share one, and WI-714's applied check
  threads that correlation.

* `constrain_vid`: a bare parameter never displaces a recorded entry and
  never contradicts one; a concrete type displaces IT. The old unification
  was a FALSE CONTRADICTION -- four vars recorded at one shared parameter,
  whose alias bound to String from the first pair, so an honest Int64 came
  back incomparable and the rule was written off as contradictory, silently
  skipping dot dispatch and stamping.

Lifting that contradiction un-suppressed a genuine type error in
examples/webots-modelling/lf1: `comm_delay_max` multiplies two Int64s and
divides by the Float baud_rate, and there is no implicit widening. Fixed by
typing byte_size/packet_size Float, with the reason recorded at the
declaration, plus the three anthill-smt-gen fixtures carrying the same spec
inline. The lf1 proofs are unchanged.

Proposal 052 "The schema T" now states the rule: a clause votes only where it
has a type, and both spellings of unconstrained contribute nothing.

Also restores the WI-714 projection doc onto `projection_columns`, which
WI-731 detached by inserting `schema_columns_tail` between them (found in
review).

Residual, filed as WI-20260819-9C2PZ: two independent `eq` calls in one rule
are still treated as correlated, so the correct citation `twoeq("a", 1)` is
refused -- verified pre-existing. It needs per-call instantiation of a
callee's type parameters.

Tests: 7 in wi741_spec_typed_column_test.rs, each fixture in its own
namespace, each with its back-out prediction stated at its site. Backing out
relation_clause_columns fails the 3 acceptance tests; backing out
constrain_vid fails the other 2; two pass either way by design and say so.

+++ rustland/anthill-smt-gen/tests/inductive_test.rs
@@ -24,7 +24,7 @@ fn inductive_kb() -> anthill_core::kb::KnowledgeBase {
 
           entity LinkParameters(
             range_max: Float, signal_speed: Float, baud_rate: Float,
-            byte_size: Int64, packet_size: Int64)
+            byte_size: Float, packet_size: Float)
           entity KinematicAssumptions(
             leader_speed_max: Float, follower_speed_max: Float,
             control_period: Float, sensor_period: Float)
@@ -84,7 +84,7 @@ fn inductive_kb() -> anthill_core::kb::KnowledgeBase {
 
           fact LinkParameters(
             range_max: 100.0, signal_speed: 300000000.0,
-            baud_rate: 1000000.0, byte_size: 8, packet_size: 32)
+            baud_rate: 1000000.0, byte_size: 8.0, packet_size: 32.0)
           fact KinematicAssumptions(
             leader_speed_max: 8.0, follower_speed_max: 8.0,
             control_period: 0.032, sensor_period: 0.008)
